brainstorming_skill/requirements.txt
openai>=1.40.0
anthropic>=0.34.0
python-dotenv>=1.0.0
networkx>=3.1
python-louvain>=0.16
requests>=2.25.0
Flask>=2.3.0
PyPDF2>=3.0.0